Completes seeding function and test
