+++ database/build.js
@@ -1,0 +1,6 @@
+var { sequelize } = require('./models.js');
+
+sequelize.sync({ force: true })
+  .then(() => {
+    sequelize.close();
+  });
