+++ db.test.js
@@ -7,8 +7,9 @@ var { PhonesCapacities, PhonesCarriers, PhonesColors } = Models;
 // this line is needed to prevent error after jest runs
 require('mysql2/node_modules/iconv-lite').encodingExists('foo');
 
+// resets the database (tears down and rebuilds)
+require('./database/build.js');
 
-// need to empty db manually first: npm run dbEmpty
 test('should add data to all tables for first record added', () => {
   db.addOne({
     name: 'string',
@@ -31,4 +32,19 @@ test('should add data to all tables for first record added', () => {
   })
 });
 
-sequelize.close();
+test('should add 100 more phone records after seed', async () => {
+  await require('./database/seed.js')
+  Promise.all([Phone.findAll(),
+        Capacity.findAll(),
+        Carrier.findAll(),
+        Color.findAll()])
+    .then(([phones, capacities, carriers, colors]) => {
+      expect(phones.length).toBe(101);
+      expect(capacities.length).toBeLessThan(10);
+      expect(carriers.length).toBeLessThan(10);
+      expect(colors.length).toBeLessThan(10);
+    })
+    .catch(() => {})
+})
+
+// sequelize.close();
