--- dbEmpty.sql
@@ -1,9 +1,0 @@
-use phoneDetails;
-
-DROP TABLE phones_capacities;
-DROP TABLE phones_carriers;
-DROP TABLE phones_colors;
-DROP TABLE capacities;
-DROP TABLE carriers;
-DROP TABLE colors;
-DROP TABLE phones;
